requirements.txt
anyio==3.6.1
APScheduler==3.9.1
arrow==1.2.3
binaryornot==0.4.4
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
chevron==0.14.0
click==8.1.3
colorama==0.4.5
cookiecutter==1.7.3
fastapi==0.79.1
h11==0.12.0
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.0
idna==3.3
Jinja2==3.1.2
jinja2-time==0.2.0
loguru==0.6.0
MarkupSafe==2.1.1
msgpack==1.0.4
multidict==6.0.2
nb-cli==0.6.7
nonebot-adapter-onebot==2.1.3
nonebot-plugin-apscheduler==0.2.0
nonebot-plugin-gocqhttp==0.6.2
nonebot2==2.0.0rc1
Pillow==9.2.0
poyo==0.5.0
prompt-toolkit==3.0.31
psutil==5.9.2
py-cpuinfo==8.0.0
pydantic==1.9.2
pyfiglet==0.8.post1
pygtrie==2.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==6.1.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
text-unidecode==1.3
tomlkit==0.10.2
typing_extensions==4.4.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.9
uvicorn==0.18.3
watchfiles==0.17.0
wcwidth==0.2.5
websockets==10.3
win32-setctime==1.1.0
yarl==1.8.1

playwright~=1.31.0
config~=0.5.1
Flask~=2.2.2